rename package mem => pstoremem.
