+++ pstoreds/peerstore.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ipfs/go-datastore"
 
 	pstore "github.com/libp2p/go-libp2p-peerstore"
-	"github.com/libp2p/go-libp2p-peerstore/mem"
+	"github.com/libp2p/go-libp2p-peerstore/pstoremem"
 )
 
 // NewPeerstore creates a peerstore backed by the provided persistent datastore.
@@ -17,6 +17,6 @@ func NewPeerstore(ctx context.Context, ds datastore.Batching) (pstore.Peerstore,
 		return nil, err
 	}
 
-	ps := pstore.NewPeerstoreWith(mem.NewKeyBook(), addrBook, mem.NewPeerMetadata())
+	ps := pstore.NewPeerstoreWith(pstoremem.NewKeyBook(), addrBook, pstoremem.NewPeerMetadata())
 	return ps, nil
 }
