+++ pstoremem/inmem_test.go
@@ -1,4 +1,4 @@
-package mem
+package pstoremem
 
 import (
 	"testing"
